backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
SQLAlchemy>=2.0.25
pydantic>=2.6.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.1
email-validator>=2.1.0.post1
python-multipart>=0.0.9
python-json-logger>=2.0.7
sentry-sdk>=2.8.0
